ref: Refactor test that writes thousands of references
One of the tests for the `FindAllBranchNames()` RPC exercises what
happens when we return thousands of references. To do so it creates a
git-update-ref(1) instance and queues up all updates to it. While this
is as efficient as it gets while using Git, it still is comparatively
expensive to do it this way as it requires locking and whatnot.

Refactor the test to just manually write out a `packed-refs` file.