operations: Merge some of TestServer_UserRevert* tests
There are a few TestServer_UserRevert* tests which do similar things,
but are outdated and use seeded repos. In preparation of the upcoming
additions to UserRevert, let's merge these tests.

This makes it easy to add new tests and also no longer uses seeded
repos, making it easier to also migrate to SHA256.