git: Drop `SubSubCmd` structure
In the `git` package we have both a `SubCmd` and a `SubSubCmd` structure
which both implement the `Cmd` interface. This abstraction is arguably
not worth it: the only thing that is different between both is that the
`SubSubCmd` also has an associated "action", like it is present in `git
remote set-url`. We thus duplicate quite a bunch of logic only just to
be able to handle this additional action.

Let's drop this abstraction and instead just add the `Action` field to
the `git.SubCmd` structure. If it's set we'll add it to the computed
arguments. If unset, we retain the current behaviour.