repository: Refactor RepackIncremental test to not assert mtimes
One of the tests for RepackIncremental uses mtimes of packfiles as a
proxy to test whether the RPC call did what it was supposed to do. This
is quite a weak proxy though for what we actually care for, which is
whether we have created a new packfile.

Refactor the test to stop checking for mtimes and instead use the
packfile count.