cgroups: Fix cyclic dependency with command package
The command package and the cgroups package have a cyclic dependency:

    - The `AddCommand()` function of the `cgroups.Manager` receives a
      `command.Command` as input.

    - The `command` package needs to have a `cgroups.Manager` so that it
      can add the command to it.

The cylic dependency is currently solved by using an interface in the
`command` package that only has the `AddCommand()` function. But it
would be preferable to just not have this cyclic dependency at all.

Refactor the code to instead accept an `exec.Cmd` in `AddCommand()` to
break the cyclic dependency.