cgroups: Stop exercising whether the `command` package logs data
One of the cgroups tests that verifies whether metrics are exposed as
expected also verifies that we get a log message. This log message is
not generated by the cgroups package though, but it's generated by the
command package. So it really is not the business of the cgroups tests
to verify that we see this log message.

Stop asserting the log message. This change also prepares us for a fix
to a cyclic dependency between the command and cgroups package.