cgroups: Remove repository-specific logic
The cgroups manager accepts a `repository.GitRepo` as argument, which is
used to compute a cgroup key that will always be the same for multiple
different Git commands as long as they're executed in the same repo. As
a result we will always put them into the same cgroup bucket.

Having this knowledge in the cgroups manager is a layering violation
though: the manager really shouldn't know about repositories at all, but
only care about commands and the bucket they're put into.

Introduce a new option `WithCgroupKey()` that allows callers to override
the cgroup key used to derive the bucket. Like this, we can pass through
the per-repository cgroup key from the Git command factory through the
command package into the cgroups manager without anybody but the Git
command factory actually knowing about repositories.