repository: Manually seed repo in CreateRepositoryFromURL test
The test that verifies CreateRepositoryFromURL's behaviour with a bunch
of optional parameters is the last remnant that still uses our beloved
"gitlab-test-mirror.git" test repository. And as we're converting our
test setups to generate test data at runtime instead of using pre-seeded
repositories we're not going to add any other callsites that need this
seed repository.

Refactor the test to generate the test data at runtime. What we really
care about in this context is a repository that has references that
would not be cloned by default to verify that we indeed clone them when
the `Mirror` parameter is set to `true`. So let's generate the data like
that and get rid of the now-unused helper function that sets up the old
"gitlab-test-mirror.git" repository seed.