housekeeping: Reinitialize test variable
In TestPruneIfNeeded, we don't reinitialize the test variable, this can
cause problems to go
undetected (https://gist.github.com/posener/92a55c4cd441fc5e5e85f27bca008721).

Fix the test and reinitialize the variable, do note that this change
removes the only test which was checking for `pruned_objects: success`.
So add a new test which creates a lot of looseObjects and also sets
their time to before the StaleObjectsGracePeriod.