structerr: Provide function to check for error equivalence
In the coordinator we have one callsite where we verify that a wrapped
error `errors.Is()` another error created via `helper.Err*f()`. This
checks works because the `grpc/status.Error` type implements `Is()`.
It stops working though with the new `structerr` package because we
don't yet implement that type, and simply checking for equality does not
work.

Implement a new `structerr.Error.Is()` function to support this usecase.