repository: Move function to transactionally create repositories
We're about to add another caller for the `createRepository()` helper
functions so that we can also transactionally and atomically create
object pools. Let's thus move the function into its own package
`repoutil`.

Note that I'm not particularly happy with the resulting package
structure. But this is service-level functionality, which means that
moving it into `internal/git` is not an option. So let's not dwell too
long over it.