repoutil: Refactor assembly of options to be more extensible
All the options that can be passed to `Create()` can be represented as
simple `git.Option`s, which is why we a `CreateOption` is defined as
appending to an array of `git.Option`s. We're about to add another
option though that doesn't fit into this model.

Refactor the code to introduce a `createConfig` structure that for now
only has the array of Git options.