objectpool: Tighten argument checks when creating object pools
We don't thoroughly verify arguments passed to the `CreateObjectPool()`
RPC call. As a result, in some scenarios we fail with not-well-defined
semantics while already busy creating the repository. This creates
problems when we refactor the code to create repositories atomically as
the exact way how we fail would be different.

Add more preliminary checks to verify that arguments are what we expect
them to look like.