git/objectpool: Resolve object pool paths when creating the structure
While we already try to resolve the storage path right now and thus fail
with an error in case the storage is not known, we don't do this for the
object pool path yet. While we check whether the object pool path is a
valid path at a later point anyway, we will require the path before
doing that check in a subsequent commit. As a consequence, the error
returned would fail in one of our tests.

Refactor the code to ask for the repository path early on so that we
don't need to change tests when introducing new logic. The new behaviour
could be claimed to be better than before as we now return a proper
error in case the path is misformed. But the previous behaviour was just
fine, too.