repoutil: Add option to skip initializing repository in Create
The `Create()` helper will make sure to call the repository-seeding
function with an already initialized Git repository so that we can skip
the step of having to call git-init(1) at all the callsites. While this
is generally useful, some of the users of `Create()` need to create the
target repository by calling git-clone(1), which isn't happy in case the
target directory exists and is not an empty directory. These callers
thus manually `os.RemoveAll()` the target directory right now before
doing the clone.

Provide a new option `WithSkipInit()` that allows callers to skip the
call to git-init(1). Convert callsites to use the new option.