localrepo: Remove build restriction for remote extra tests
When converting the localrepo package to support SHA256 as object format
we couldn't yet convert the remote extra tests. This was because they
relied on creating the repository via the `CreateRepository` RPC, but
the RPC didn't yet support creating SHA256 repositories.

This has since been implemented, so let's enable these tests.