gittest: Simplify setup of object format
We introduced the ability to use `gittest.CreateRepository()` back when
the `git.ObjectHash` didn't yet have a `Format` field. We thus have a
separate field that tracks the extra arguments required for git-init(1)
to set up the repository with the expected object format.

This is not required anymore as we can now just reuse the `Format`, so
let's simplify the code.