housekeeping: Enable testing optimization strategy with SHA256
Enable testing the optimization strategy with SHA256. While we had the
logic in place to make object sizes conditional on the object hash, we
have since started to write empty trees in `gittest.WriteCommit()`. As a
result the object size has increased.

Fix these stale object sizes and enable testing with SHA256.