status: Return ResourceExhausted code if no more file descriptors
When Gitaly has troubles with acquiring more file descriptors from the
system it returns Internal or Unknown status codes back to the caller.
That doesn't tell the caller anything meaningful, so the caller re-
attempts the operation. Re-attempts create additional load on the
Gitaly. What we want instead is a backoff that will allow Gitaly to
handle in process requests and lower file descriptors usage. That is
why now Gitaly returns ResourceExhausted status code in case it
encounters a problem with file descriptors.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/4688