command: Test the 'too many open files' error is returned
We have statushandler middleware that changes status code of the RPC
depending on different conditions. One of such conditions is the
syscall.EMFILE error. This test is added to make sure this particular
error is returned when there is no more file descriptors available.
Otherwise, the statushandler middleware needs to be adjusted to catch
the proper error.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/4688