namespace: Simplify test setup
Simplify the test setup for the `RenameNamespace()` RPC. While we set up
the logic to iterate over multiple testcases and execute some subset of
the verification conditionally on the testcase, we only have a single
testcase anyway.

I don't expect that we'll extend the namespace service at any point in
time anymore as we'd rather want to remove it altogether. So let's
refactor the test to not have the boilerplate anymore.