operations: Merge some of Test*UserUpdateSubmodule* tests
There are a few Test*UserUpdateSubmodule* tests which do similar things,
but are outdated and use seeded repos. In preparation of the upcoming
additions to UserUpdateSubmodule, let's merge these tests.

This makes it easy to add new tests and also no longer uses seeded
repos, making it easier to also migrate to SHA256.