prometheus: Convert to use LabKit's `WithGoBuildInformation()` option
Starting with LabKit v1.17.0 there is a new `WithGoBuildInformation()`
option that uses the `debug.BuildInfo` structure provided by Go to
derive information about the executable. This new infrastructure
replaces `WithBuildInformation()` and exposes additional information
that was not previously exposed:

    - The Go version we're running at.

    - Whether sources have been modified.

    - The module path.

    - The exact commit ID that was built.

Convert Gitaly, Praefect and our blackbox to use this new build option
instead of the old infrastructure. Note that while we only conditionally
pass the option in case `ReadBuildInfo()` returns `true`, Go always
embeds the build information into binaries when building with modules
support enabled. So in practice it should always be present.