structerr: Add option to override gRPC error code
The `structerr` package propagates error codes of wrapped errors so that
we by default retain the most specific one. While this works in general,
there may be cases where we explicitly want to override the error code
with a different one.

Add a new option `WithGRPCCode()` that allows us to do so.