Merge backchannel client handshaker constructors
Backchannel has currently two constructors, one taking in a config
and the other one not. This commit merges the constructors into one
which takes in a custom configuration type. This hides the Yamux
details from outside the package as some of the configuration options
are overridden by backchannel itself. The current configuration is
kept so there should be no behavioral differences.