repository: Add `^{commit}` to `ResolveRevision`
Currently the tests in `TestFetchSourceBranch` use `ResovleRevision` to
check if the ref exists. While this is correct, this only checks that
the ref file exists, it doesn't verify if the underlying object exists
in git. We want to be more careful here, because we want to use
quarantined repository for `FetchSourceBranch` in the following commit.

With quarantined repositories, while the objects/packfiles are present
in the quarantined folder, ref updates take place in the main folder. So
these tests could provide us false positive results, since the refs
would be updated with the provided commit IDs, but the commit object
themselves might not be present. Let's avoid such issues and be more
explicit in these tests.