repository: Use quarantined repos in `FetchSourceBranch`
Currently `FetchSourceBranch` uses the following flow of operation:
1. Resolve the objects to be added
2. Check if target repository already contains required objects
3. Fetch the objects required to the target repository
4. Update the ref

Since we don't use quarantined repositories here, we might leave the
target repositories corrupted if we fail between 3-4. This is
problematic in nature. While the ideal solution is to patch
Git (https://gitlab.com/gitlab-org/git/-/issues/134) to use a
quarantine directory, this could take some time to implement. So for now
we're going to use a quarantined repository via Gitaly.

Since this is a high risk feature, we add it behind the newly introduced
`FetchSourceBranchQuarantined` feature flag.