Use Yamux configuration in Praefect by default
`praefect_use_yamux_configuration_for_gitaly` feature flag enables
Praefect to read Yamux configuration from the config file. This
commit removes the feature flag and enables the new default settings
that were tested behind it. We're skipping default enabling the flag
as it requires two connections sets and the feature flagged connections
are not used for all of the functionality like elections. To avoid this
complexity, we'll just remove the flag directly and enable the new
defaults.

The new defaults lower the receive buffer size in Yamux leading to less
memory usage but possibly worse throughput in high latency scenarios.
We were previously using custom settings but are now returning to the
Yamux defaults as they are also used in the clients for the sidechannels.
This ensures Praefect doesn't buffer excessively with sidechannels in use.

Changelog: changed