localrepo: Teach UpdateRef to accept options
It's useful to be able to pass options to UpdatRef to modify the
behavior of the git-update-ref(1) call. For instance, there are times
when we need to set GIT_OBJECT_DIRECTORY and
GIT_ALTERNATE_OBJECT_DIRECTORIES. Introcue a new UpdateRefOption with
the first option, WithEnv that allows callers to set environment
variables that get passed through to the git command.