commit: Adds max batch size per file type
Why?

https://gitlab.com/gitlab-org/gitlab/-/issues/17698

On 14.5 we added improved diffs for Python Notebook (ipynbs). These
are larges that have embedded images, and the limit of 512kb makes
blocks the improved diffs from being displayed in the majority of
cases. But we don't want to increase the diff size for all file
types, since that could lead to performance degradation, only to
files that really need it. The size of the diff will be controlled
from the consumer side.

What?

Adds a new option to the CommitDiffRequest that allows users to
specify a different maximum value per file type. This value will
override MaxPatchBytes and SafeMaxPatchBytes, but not MaxBytes or
SafeMaxBytes.

Changelog: added