localrepo: Add WriteTree method
Add a WriteTree and WriteTestTree methods. These will eventually serve
as the unified way to create trees in a git repository. Currently there
is also the internal/git/gittest package which has a WriteTree method.
This method is considered deprecated and should no longer be used in the
future.

A subsequent commit will refactor almost all test code to use the
localrepo package for writing trees.