catfile: fix another 64-bit alignment error on objectInfoReader
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5223 fixed one
64-bit alignment error on Raspberry Pi for `TestCache_ObjectReader`,
but `TestCache_ObjectInfoReader` also failed due to an alignment
error.

We need to reorder the fields to ensure the atomic values are 64-bit
aligned.

We also need to move `queueCounters` to the top since this appears to
be necessary for GetTreeEntries as well.

Once we switch to Go 1.19, we can use the new atomic types
to avoid this issue entirely:
https://gitlab.com/gitlab-org/gitaly/-/issues/4702

Changelog: fixed