Implement service config generator
The method config is generated from the list of read-only RPCs. Mutator
RPCs are likely not safe for auto-retry. Some upsert RPCs are built
with idempotency in mind, but we cannot assume that would be the case
for all existing RPCs and future RPCs. As a result, upper layers are
responsible for mutator RPCs retry.

The default load balancing policy is set to round-robin.