Add an integration test to verify auto-retry policy
The test starts a so-called Gitaly server serving an Unary call. Client
dials with different dialers and enabled WithDefaultLoadBalancing dial
option. Unavailable and DeadlineExceeded errors should invoke auto-retry
mechanism so that client retry the RPC until reaching the upper limit.

Due to the difficulties when setting up tests, it's not feasible to
verify the round-robin behaviors at this point.