repository: Fix intermingled stdout/stderr output for git-fsck(1)
We use two different buffers when spawning git-fsck(1) for its stdout
and stderr. This is inefficient as we now need to handle the overhead
for two buffers and need to spawn two Goroutines for them. Furthermore,
if git-fsck(1) was to print alternately to stdout and stderr, the order
of these messages gets lost as we used to just concatenate stdout and
stderr.

Fix this by using a single buffer for both standard streams. While at
it, convert the code to use a `strings.Builder` instead of a
`bytes.Buffer`, which is more efficient.