repository: Disable misleading dangling checks in git-fsck(1)
The `Fsck` RPC uses git-fsck(1) to inform the user about repository
corruption. In case the Git command returns with a non-zero exit code
the RPC will return both its stdout and stderr to the caller via the
`Error` field that is part of the `FsckResponse`.

By default, git-fsck(1) will not only print information about repository
corruption though, but also about dangling objects that aren't reachable
via any reference. But as it is totally expected that repositories have
dangling objects, it is only distracting from the actual error messages
that the repository owner would care about.

Stop git-fsck(1) from checking for dangling objects to improve the
signal to noise ratio. While at it, let's also explicitly ask it to not
print any progress -- while it wouldn't do that anyway because it is not
printing to a TTY, it stops the next reader of the code from wondering
whether it would ever do that.

Changelog: improved