build-proto-gem: Add `--skip-verify-tag` option
The script to build the Gitaly ruby gem compares the VERSION file with
the current tag (via `git describe --tags`), this is a safety check so
that we don't build an publish non-tagged releases.

But this also blocks us from building the gem for local testing. So add
a new option `---skip-verify-tag` to the build script which can skip
this verification.

Since we're adding OptionParser also move the previously used argument
to an option.