limithandler: Fix flaky test caused by stream being closed async
When rate-limiting streams, the limiting only kicks in upon receiving
the first Protobuf message so that we can derive the limiting key based
on some parameters. This means that the client can already start sending
requests even though the server side has not yet decided whether it
wants to allow the RPC call or rate-limit it.

This causes one of our tests for a full-duplex call to be flaky as we
sometimes see an `io.EOF` when sending the requests. This is caused by
us sending multiple requests to the server: when it has received the
first request and closes the stream before we have sent all 10 of our
requests, then the client will indeed get an early `io.EOF`. This is
entirely expected though due to the async nature of the rate-limiting
for full-duplex calls.

Fix the test so that it gracefully handles an early `io.EOF`.