benchmarking: Add benchmarking script for client
Using `ghz` requires a large number of parameters which would be
quite painful to write in YAML. Create a wrapper script for the client
host that does some basic parameter verification and then invokes `ghz`.

Currently the --concurrency[0] and --rps[1] values used are arbitrary.
In the future we should look into configuring these appropriately per
RPC. For example, 100 `OptimizeRepository` requests per second is not a
useful scenario.

[0] https://ghz.sh/docs/options#-c---concurrency
[1] https://ghz.sh/docs/options#-r---rps