benchmarking: Add profiling script
Understanding where Gitaly and Git are spending their time, as well as
general system health are critical to useful benchmarking. Add a script
to the Gitaly node to run `perf` and a number of `libbpf-tools`
utilities while the node is under load.

Running this introduces a performance overhead of ~10%, mostly from
`perf`, which is run twice simultaneously. Once to profile only Gitaly
using `--call-graph=fp`, which works well with Golang, and again for the
system as whole using `--call-graph=dwarf`, which is more accurate for
Git and other C programs. The DWARF output is ~10x larger than function
pointer, causing flamegraphs built from it to take proportionately
longer, typically longer than the duration profiled.

The `libbpf-tools` utilities used are a bit of a grab bag, but quite
lightweight to run. This are BPF CO-RE utilities that run much more
lightly than `bcc`, which can be a resource hog. These focus primarily
on determing the amount of delay block I/O imposes, which may be useful
in determining how much of a penalty slower storage imposes on Gitaly.
Currently the only RPC being tested is `FindCommit`, which being
read-only hits the kernel page cache 100% of the time after the first
request.

- biolatency: Histogram of the latency of block I/O operations for each
  attached disk.
  https://github.com/iovisor/bcc/blob/master/tools/biolatency_example.txt

- biotop: List of processes performing the most block I/O.
  https://github.com/iovisor/bcc/blob/master/tools/biotop_example.txt

- execsnoop: List of all processes forked by Gitaly and their
  arguments.
  https://github.com/iovisor/bcc/blob/master/tools/execsnoop_example.txt

- cpudist: Histogram of durations that programs executed by the kernel,
  or with the `--offcpu` flag, how long they were slept.
  https://github.com/iovisor/bcc/blob/master/tools/cpudist_example.txt

- cachestat: Statistics regarding kernel page cache hit rate.
  https://github.com/iovisor/bcc/blob/master/tools/cachestat_example.txt

Note that the links above are to the `bcc` documentation for each tool
used. The arguments the `bcc` version takes may vary a bit from what
`libbpf-tools` allows, but they perform the same task.

Further work is needed for this be fully useable, most notably tracking
CPU and memory utilization. This is difficult with polling tools like
Prometheus's `node-exporter`, as most of the system load is typically
from short-lived Git processes that may spawn and exit between polling
intervals.